requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
sqlmodel
python-dotenv
google-genai
